Split the error-injection driver into a mocks subfolder.
Review feedback on #2169: one 1061-line mock_error_agent.cpp was doing too many
jobs at once. Most of its bulk was not the driver at all, but a 280-line
scenario table declared inside main, plus the mock engine, the injection switch,
and the output formatting.

Move all of it to test/gtest/mocks/error_injection/ and separate the pieces by
what a reader is likely to come for:

  error_injection.h    the model: injection sites, actions, expected agent
                       behavior, the scenario and observation structs, and the
                       mock engine and agent declarations
  mock_error_engine.cpp  the engine, the mock agent, and applyInjection
  error_scenarios.cpp  the table, lifted out of main behind errorScenarios(),
                       with the note on why these 46 cases and not the full
                       site-by-status matrix
  scenario_report.cpp  column widths, verdict wording, table and csv output
  mock_error_agent.cpp what is left: the northbound sequence and main

The anonymous namespace had to become a named one to span translation units, so
the model now lives in mocks::error_injection.

This is a pure move: no scenario, status, or expectation changed. Verified by
diffing the table, --csv, and --list output against the same binary before the
split, all three byte-identical, still 46 scenarios with 0 failures, and the
executable target name and path are unchanged so existing run instructions hold.
