+++ test/gtest/mocks/error_injection/error_scenarios.cpp
@@ -1,0 +1,332 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
+ * The scenario table. Twenty status-returning southbound calls against the
+ * status codes a plugin may return would be several hundred single-fault runs,
+ * nearly all of which exercise the same few lines of agent mapping code, so the
+ * set below is cut down to the cases that can differ: every site is covered at
+ * least once, with a second status wherever the agent's mapping is not uniform
+ * (a site that collapses to NIXL_ERR_BACKEND gets both NIXL_ERR_BACKEND and
+ * something else, so the collapse is visible) and extra entries where the agent
+ * branches on the value itself, such as NIXL_ERR_REMOTE_DISCONNECT,
+ * NIXL_ERR_NOT_SUPPORTED, and NIXL_IN_PROG.
+ *
+ * Grouped by the northbound call under test: creation, registration, metadata,
+ * connection, transfer, memory views, notifications, optional APIs, cleanup,
+ * and agent-side rejections that need no injection at all.
+ */
+
+#include "mocks/error_injection/error_injection.h"
+
+namespace mocks::error_injection {
+
+const std::vector<scenario> &
+errorScenarios() {
+    static const std::vector<scenario> scenarios{
+        {"baseline",
+         action_t::COMPLETE,
+         injection_site_t::NONE,
+         NIXL_SUCCESS,
+         NIXL_SUCCESS,
+         behavior_t::COMPLETED},
+
+        {"create.getConnInfo.invalidParam",
+         action_t::CREATE_BACKEND,
+         injection_site_t::GET_CONN_INFO,
+         NIXL_ERR_INVALID_PARAM,
+         NIXL_ERR_INVALID_PARAM,
+         behavior_t::PASS_THROUGH},
+        {"create.getConnInfo.backend",
+         action_t::CREATE_BACKEND,
+         injection_site_t::GET_CONN_INFO,
+         NIXL_ERR_BACKEND,
+         NIXL_ERR_BACKEND,
+         behavior_t::PASS_THROUGH},
+        {"create.connect.invalidParam",
+         action_t::CREATE_BACKEND,
+         injection_site_t::CONNECT,
+         NIXL_ERR_INVALID_PARAM,
+         NIXL_ERR_INVALID_PARAM,
+         behavior_t::PASS_THROUGH},
+        {"create.connect.backend",
+         action_t::CREATE_BACKEND,
+         injection_site_t::CONNECT,
+         NIXL_ERR_BACKEND,
+         NIXL_ERR_BACKEND,
+         behavior_t::PASS_THROUGH},
+
+        {"register.registerMem.invalidParam",
+         action_t::REGISTER_MEM,
+         injection_site_t::REGISTER_MEM,
+         NIXL_ERR_INVALID_PARAM,
+         NIXL_ERR_BACKEND,
+         behavior_t::COLLAPSED},
+        {"register.registerMem.notFound",
+         action_t::REGISTER_MEM,
+         injection_site_t::REGISTER_MEM,
+         NIXL_ERR_NOT_FOUND,
+         NIXL_ERR_BACKEND,
+         behavior_t::COLLAPSED},
+        {"register.registerMem.backend",
+         action_t::REGISTER_MEM,
+         injection_site_t::REGISTER_MEM,
+         NIXL_ERR_BACKEND,
+         NIXL_ERR_BACKEND,
+         behavior_t::COLLAPSED},
+        {"register.loadLocalMD.invalidParam",
+         action_t::REGISTER_MEM,
+         injection_site_t::LOAD_LOCAL_MD,
+         NIXL_ERR_INVALID_PARAM,
+         NIXL_ERR_BACKEND,
+         behavior_t::COLLAPSED},
+        {"register.loadLocalMD.backend",
+         action_t::REGISTER_MEM,
+         injection_site_t::LOAD_LOCAL_MD,
+         NIXL_ERR_BACKEND,
+         NIXL_ERR_BACKEND,
+         behavior_t::COLLAPSED},
+        {"register.getPublicData.invalidParam",
+         action_t::REGISTER_MEM,
+         injection_site_t::GET_PUBLIC_DATA,
+         NIXL_ERR_INVALID_PARAM,
+         NIXL_ERR_BACKEND,
+         behavior_t::COLLAPSED},
+        {"register.getPublicData.backend",
+         action_t::REGISTER_MEM,
+         injection_site_t::GET_PUBLIC_DATA,
+         NIXL_ERR_BACKEND,
+         NIXL_ERR_BACKEND,
+         behavior_t::COLLAPSED},
+
+        {"metadata.loadRemoteConnInfo.invalidParam",
+         action_t::LOAD_REMOTE_MD,
+         injection_site_t::LOAD_REMOTE_CONN_INFO,
+         NIXL_ERR_INVALID_PARAM,
+         NIXL_ERR_INVALID_PARAM,
+         behavior_t::PASS_THROUGH},
+        {"metadata.loadRemoteConnInfo.notAllowed",
+         action_t::LOAD_REMOTE_MD,
+         injection_site_t::LOAD_REMOTE_CONN_INFO,
+         NIXL_ERR_NOT_ALLOWED,
+         NIXL_ERR_NOT_ALLOWED,
+         behavior_t::PASS_THROUGH},
+        {"metadata.loadRemoteConnInfo.notSupported",
+         action_t::LOAD_REMOTE_MD,
+         injection_site_t::LOAD_REMOTE_CONN_INFO,
+         NIXL_ERR_NOT_SUPPORTED,
+         NIXL_ERR_BACKEND,
+         behavior_t::COLLAPSED},
+        {"metadata.loadRemoteMD.invalidParam",
+         action_t::LOAD_REMOTE_MD,
+         injection_site_t::LOAD_REMOTE_MD,
+         NIXL_ERR_INVALID_PARAM,
+         NIXL_ERR_INVALID_PARAM,
+         behavior_t::PASS_THROUGH},
+        {"metadata.loadRemoteMD.notFound",
+         action_t::LOAD_REMOTE_MD,
+         injection_site_t::LOAD_REMOTE_MD,
+         NIXL_ERR_NOT_FOUND,
+         NIXL_ERR_NOT_FOUND,
+         behavior_t::PASS_THROUGH},
+
+        {"connection.connect.backend",
+         action_t::MAKE_CONNECTION,
+         injection_site_t::CONNECT,
+         NIXL_ERR_BACKEND,
+         NIXL_ERR_BACKEND,
+         behavior_t::PASS_THROUGH},
+        {"connection.connect.notFound",
+         action_t::MAKE_CONNECTION,
+         injection_site_t::CONNECT,
+         NIXL_ERR_NOT_FOUND,
+         NIXL_ERR_NOT_FOUND,
+         behavior_t::PASS_THROUGH},
+
+        {"transfer.prepXfer.invalidParam",
+         action_t::CREATE_XFER,
+         injection_site_t::PREP_XFER,
+         NIXL_ERR_INVALID_PARAM,
+         NIXL_ERR_INVALID_PARAM,
+         behavior_t::PASS_THROUGH},
+        {"transfer.prepXfer.notFound",
+         action_t::CREATE_XFER,
+         injection_site_t::PREP_XFER,
+         NIXL_ERR_NOT_FOUND,
+         NIXL_ERR_NOT_FOUND,
+         behavior_t::PASS_THROUGH},
+        {"transfer.postXfer.backend",
+         action_t::POST_XFER,
+         injection_site_t::POST_XFER,
+         NIXL_ERR_BACKEND,
+         NIXL_ERR_BACKEND,
+         behavior_t::PASS_THROUGH},
+        {"transfer.postXfer.remoteDisconnect",
+         action_t::POST_XFER,
+         injection_site_t::POST_XFER,
+         NIXL_ERR_REMOTE_DISCONNECT,
+         NIXL_ERR_REMOTE_DISCONNECT,
+         behavior_t::PASS_THROUGH_RECOVERABLE},
+        {"transfer.checkXfer.backend",
+         action_t::CHECK_XFER,
+         injection_site_t::CHECK_XFER,
+         NIXL_ERR_BACKEND,
+         NIXL_ERR_BACKEND,
+         behavior_t::PASS_THROUGH},
+        {"transfer.checkXfer.remoteDisconnect",
+         action_t::CHECK_XFER,
+         injection_site_t::CHECK_XFER,
+         NIXL_ERR_REMOTE_DISCONNECT,
+         NIXL_ERR_REMOTE_DISCONNECT,
+         behavior_t::PASS_THROUGH_RECOVERABLE},
+        {"transfer.checkXfer.canceled",
+         action_t::CHECK_XFER,
+         injection_site_t::CHECK_XFER,
+         NIXL_ERR_CANCELED,
+         NIXL_ERR_CANCELED,
+         behavior_t::PASS_THROUGH},
+        {"transfer.releaseReqH.backend",
+         action_t::RELEASE_XFER,
+         injection_site_t::RELEASE_REQ,
+         NIXL_ERR_BACKEND,
+         NIXL_ERR_REPOST_ACTIVE,
+         behavior_t::TRANSFORMED},
+
+        {"memView.remote.notSupported",
+         action_t::PREP_REMOTE_MEM_VIEW,
+         injection_site_t::PREP_REMOTE_MEM_VIEW,
+         NIXL_ERR_NOT_SUPPORTED,
+         NIXL_ERR_NOT_SUPPORTED,
+         behavior_t::PASS_THROUGH},
+        {"memView.remote.backend",
+         action_t::PREP_REMOTE_MEM_VIEW,
+         injection_site_t::PREP_REMOTE_MEM_VIEW,
+         NIXL_ERR_BACKEND,
+         NIXL_ERR_BACKEND,
+         behavior_t::PASS_THROUGH},
+        {"memView.local.notSupported",
+         action_t::PREP_LOCAL_MEM_VIEW,
+         injection_site_t::PREP_LOCAL_MEM_VIEW,
+         NIXL_ERR_NOT_SUPPORTED,
+         NIXL_ERR_NOT_SUPPORTED,
+         behavior_t::PASS_THROUGH},
+        {"memView.local.backend",
+         action_t::PREP_LOCAL_MEM_VIEW,
+         injection_site_t::PREP_LOCAL_MEM_VIEW,
+         NIXL_ERR_BACKEND,
+         NIXL_ERR_BACKEND,
+         behavior_t::PASS_THROUGH},
+
+        {"notification.getNotifs.backend",
+         action_t::GET_NOTIFS,
+         injection_site_t::GET_NOTIFS,
+         NIXL_ERR_BACKEND,
+         NIXL_ERR_BACKEND,
+         behavior_t::PASS_THROUGH},
+        {"notification.getNotifs.invalidParam",
+         action_t::GET_NOTIFS,
+         injection_site_t::GET_NOTIFS,
+         NIXL_ERR_INVALID_PARAM,
+         NIXL_ERR_INVALID_PARAM,
+         behavior_t::PASS_THROUGH},
+        {"notification.genNotif.backend",
+         action_t::GEN_NOTIF,
+         injection_site_t::GEN_NOTIF,
+         NIXL_ERR_BACKEND,
+         NIXL_ERR_BACKEND,
+         behavior_t::PASS_THROUGH},
+        {"notification.genNotif.notFound",
+         action_t::GEN_NOTIF,
+         injection_site_t::GEN_NOTIF,
+         NIXL_ERR_NOT_FOUND,
+         NIXL_ERR_NOT_FOUND,
+         behavior_t::PASS_THROUGH},
+
+        {"optional.queryMem.notSupported",
+         action_t::QUERY_MEM,
+         injection_site_t::QUERY_MEM,
+         NIXL_ERR_NOT_SUPPORTED,
+         NIXL_ERR_NOT_SUPPORTED,
+         behavior_t::PASS_THROUGH},
+        {"optional.queryMem.invalidParam",
+         action_t::QUERY_MEM,
+         injection_site_t::QUERY_MEM,
+         NIXL_ERR_INVALID_PARAM,
+         NIXL_ERR_INVALID_PARAM,
+         behavior_t::PASS_THROUGH},
+        {"optional.estimateXferCost.notSupported",
+         action_t::ESTIMATE_XFER_COST,
+         injection_site_t::ESTIMATE_XFER_COST,
+         NIXL_ERR_NOT_SUPPORTED,
+         NIXL_ERR_NOT_SUPPORTED,
+         behavior_t::PASS_THROUGH},
+        {"optional.estimateXferCost.mismatch",
+         action_t::ESTIMATE_XFER_COST,
+         injection_site_t::ESTIMATE_XFER_COST,
+         NIXL_ERR_MISMATCH,
+         NIXL_ERR_MISMATCH,
+         behavior_t::PASS_THROUGH},
+
+        {"cleanup.deregisterMem.backend",
+         action_t::DEREGISTER_MEM,
+         injection_site_t::DEREGISTER_MEM,
+         NIXL_ERR_BACKEND,
+         NIXL_SUCCESS,
+         behavior_t::IGNORED},
+        {"cleanup.unloadMD.backend",
+         action_t::DEREGISTER_MEM,
+         injection_site_t::UNLOAD_MD,
+         NIXL_ERR_BACKEND,
+         NIXL_SUCCESS,
+         behavior_t::IGNORED},
+        {"cleanup.disconnect.backend",
+         action_t::INVALIDATE_REMOTE_MD,
+         injection_site_t::DISCONNECT,
+         NIXL_ERR_BACKEND,
+         NIXL_SUCCESS,
+         behavior_t::IGNORED},
+
+        {"control.pollToCompletion",
+         action_t::POLL_TO_COMPLETION,
+         injection_site_t::POST_XFER,
+         NIXL_IN_PROG,
+         NIXL_SUCCESS,
+         behavior_t::PROGRESSED},
+        {"control.repostActive",
+         action_t::REPOST_ACTIVE,
+         injection_site_t::POST_XFER,
+         NIXL_IN_PROG,
+         NIXL_ERR_REPOST_ACTIVE,
+         behavior_t::AGENT_GENERATED},
+        {"control.statusBeforePost",
+         action_t::STATUS_BEFORE_POST,
+         injection_site_t::NONE,
+         NIXL_SUCCESS,
+         NIXL_ERR_NOT_POSTED,
+         behavior_t::AGENT_GENERATED},
+        {"control.telemetryDisabled",
+         action_t::TELEMETRY_DISABLED,
+         injection_site_t::NONE,
+         NIXL_SUCCESS,
+         NIXL_ERR_NO_TELEMETRY,
+         behavior_t::AGENT_GENERATED},
+    };
+
+    return scenarios;
+}
+
+} // namespace mocks::error_injection
